Fix conditional cost button
